+++ client/src/app/util/util.module.ts
@@ -19,17 +19,20 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import {SpinnerComponent} from './spinner/spinner.component';
+import { MomentDatePipe } from './moment-date.pipe';
 
 
 @NgModule({
   declarations: [
-    SpinnerComponent
+    SpinnerComponent,
+    MomentDatePipe
   ],
   imports: [
     CommonModule
   ],
   exports: [
     SpinnerComponent,
+    MomentDatePipe,
   ]
 })
 export class UtilModule { }